cloud-cost-dashboard/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
click==8.1.8
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
requests
fastapi-cache2
redis
requirements.txt - Project Dependencies

sqlalchemy==2.0.23
pydantic==2.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6